dtraced: Add a 'dispatched_jobs' list to prevent races.
There was a race condition in dtraced which could cause a deadlock,
however we never saw it actually trigger in practice because none of
the clients had a bug that would cause the deadlock in dtraced.

Specifically, the race condition is as follows:

 (1) Client 1 attempts to recv() from dtraced (e.g. an ELF file)
 (2) Client 1's fd gets disabled
 (3) However, the first thing in our list is Client 2's job
 (4) Client 2's job get processed (e.g. sends an ELF file)
 (5) However, the message buffer is full and send() blocks.
 (6) Client 2 never calls recv(), and a dtraced worker is permanently
     stuck in a send().

This commit fixes this by simply checking that we are dispatching events
on file descriptors that are actually attempting to recv() rather than
all jobs at once.

+++ cddl/contrib/opensolaris/cmd/dtraced/dtraced_state.c
@@ -180,6 +180,12 @@ init_state(struct dtd_state *s, int ctrlmachine, int nosha, int n_threads,
 	}
 
 	if ((err = mutex_init(
+	    &s->dispatched_jobsmtx, NULL, "joblist", CHECKOWNER_YES)) != 0) {
+		dump_errmsg("Failed to create joblist mutex: %m");
+		return (-1);
+	}
+
+	if ((err = mutex_init(
 	    &s->kill_listmtx, NULL, "kill list", CHECKOWNER_YES)) != 0) {
 		dump_errmsg("Failed to create kill list mutex: %m");
 		return (-1);
